chore: removed the schema from the redux state
